working version of eval using explicit polynomial transform

+++ SO3Fun/@SO3FunHarmonic/eval.m
@@ -1,4 +1,4 @@
-function f =  eval(F,v,varargin)
+function f =  eval(F,rot,varargin)
 % evaluates the rotational harmonic on a given set of points using NFSOFT
 % 
 % gives a align for numel(F)==1 or otherwise transform the align of v to a
@@ -15,6 +15,12 @@ function f =  eval(F,v,varargin)
 %   f - double
 %
 
+if check_option(varargin,'v2')
+  f = eval2(F,rot,varargin{:});
+  return
+end
+
+
 persistent keepPlan;
 
 % kill plan
@@ -24,19 +30,19 @@ if check_option(varargin,'killPlan')
   return
 end
 
-if isempty(v), f = []; return; end
+if isempty(rot), f = []; return; end
 
-s=size(v);
-v = v(:);
+s=size(rot);
+rot = rot(:);
 
 if F.bandwidth == 0
-  f = ones(size(v)).*F.fhat;
+  f = ones(size(rot)).*F.fhat;
   if numel(F)==1, f=reshape(f,s); end
   return;
 end
 
 % maybe we should set antipodal
-F.antipodal = check_option(varargin,'antipodal') || (isa(v,'orientation') && v.antipodal);
+F.antipodal = check_option(varargin,'antipodal') || (isa(rot,'orientation') && rot.antipodal);
 
 % extract bandwidth
 L = min(F.bandwidth,get_option(varargin,'bandwidth',inf));
@@ -53,17 +59,20 @@ if isempty(plan)
   % 2^4 -> nfsoft-represent
   % 2^2 -> nfsoft-use-DPT
   nfsoft_flags = bitor(2^4,4);
+  % nfft cutoff - 4
+  % fpt kappa - 1000
+  % fftw_size -> 2*ceil(1.5*L)
   % initialize nfsoft plan
-  plan = nfsoftmex('init',L,length(v),nfsoft_flags,0,4,1000,2*ceil(1.5*L)); 
+  plan = nfsoftmex('init',L,length(rot),nfsoft_flags,0,4,1000,2*ceil(1.5*L)); 
 
   %  set rotations in Euler angles (nodes)
-  nfsoftmex('set_x',plan,Euler(v,'nfft').');
+  nfsoftmex('set_x',plan,Euler(rot,'nfft').');
 
   % node-dependent precomputation
   nfsoftmex('precompute',plan);
 
 end
-f=zeros([length(v) size(F)]);
+f=zeros([length(rot) size(F)]);
 for k=1:length(F)
 
 % set Fourier coefficients
@@ -73,7 +82,7 @@ nfsoftmex('set_f_hat',plan,reshape(F.fhat(1:Ldim,k),[],1));
 nfsoftmex('trafo',plan);
 
 % get function values from plan
-f(:,k) = real(nfsoftmex('get_f',plan));
+f(:,k) = nfsoftmex('get_f',plan);
 
 end
